Split schema factory, add TODOs from review.
